+++ local-s3-core/src/main/java/com/robothy/s3/core/exception/VersionedObjectNotExistException.java
@@ -2,6 +2,9 @@ package com.robothy.s3.core.exception;
 
 /**
  * Versioned object not exists. Client side exception.
+ *
+ * <p>The message is the one of Amazon S3; the key and the version are reported by the {@code <Key>} and the
+ * {@code <VersionId>} of the error, where Amazon S3 reports them.
  */
 public class VersionedObjectNotExistException extends LocalS3Exception {
 
@@ -12,7 +15,8 @@ public class VersionedObjectNotExistException extends LocalS3Exception {
    * @param versionId the version ID.
    */
   public VersionedObjectNotExistException(String key, String versionId) {
-    super(S3ErrorCode.NoSuchVersion, "Object(key=" + key + ") hasn't version '" + versionId + "'.");
+    this(versionId);
+    setKey(key);
   }
 
   /**
@@ -21,7 +25,8 @@ public class VersionedObjectNotExistException extends LocalS3Exception {
    * @param version the object version.
    */
   public VersionedObjectNotExistException(String version) {
-    super(S3ErrorCode.NoSuchVersion, "The object doesn't have version ID: " + version);
+    super(S3ErrorCode.NoSuchVersion);
+    setVersionId(version);
   }
 
 }