+++ local-s3-rest/src/main/java/com/robothy/s3/rest/constants/AmzHeaderNames.java
@@ -7,6 +7,8 @@ public class AmzHeaderNames {
 
   public static final String X_AMZ_REQUEST_ID = "x-amz-request-id";
 
+  public static final String X_AMZ_ID_2 = "x-amz-id-2";
+
   public static final String X_AMZ_VERSION_ID = "x-amz-version-id";
 
   public static final String X_AMZ_CONTENT_SHA256 = "x-amz-content-sha256";